requirements.txt
dearpygui>=1.10
pillow
numpy
